Rename `constants.py` to `settings.py` in `articles` app
